+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@ Version 0.2.0
 -------------
 
 New features:
+* Simplified code for dump file processing: new helper class DumpProcessingController
 * Support for resolving site links, based on information from the sites table dump
   (as demonstrated in a new example program)
 * Support for SnakGroups (data model updated to group Snaks by property in all lists)
@@ -15,6 +16,12 @@ Bug fixes:
 * Support SomeValueSnak and NoValueSnak in references (Issue #44)
 * Use correct site links when importing data from dumps (Issue #37)
 * Do not attempt to download unfinished dump files (Issue #63)
+
+Incompatible changes:
+* The processing of dumpfiles was simplified, using a new class DumpProcessingController.
+  The former method WmfDumpFileManager#processRecentRevisionDumps() was replaced by
+  DumpProcessingController#processAllRecentRevisionDumps(). See the examples for example
+  code. 
 
 
 Version 0.1.0